+++ RELEASES.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## Development
 
+- Add a step builder for a more comprehensible way of configuring and creating `RigidBody` instances representing simple objects (such as cubes or spheres)
 - Move guide documents to a dedicated directory
 
 ## 0.2.0 (11.04.2021)